Moved all cuda code to double precision

+++ pyQCD/core/include/cuda_utils.h
@@ -17,23 +17,23 @@
 
 typedef cusp::device_memory devMem;
 typedef cusp::host_memory hostMem;
-typedef cusp::coo_matrix<int, cusp::complex<float>, hostMem> complexHybridHost;
-typedef cusp::ell_matrix<int, cusp::complex<float>, devMem> complexHybridDev;
+typedef cusp::coo_matrix<int, cusp::complex<double>, hostMem> complexHybridHost;
+typedef cusp::ell_matrix<int, cusp::complex<double>, devMem> complexHybridDev;
 
 namespace pyQCD
 {
   namespace cuda {
     void createSource(const int site[4], const int spin, const int colour,
 		      const complexHybridDev& smearingOperator,
-		      cusp::array1d<cusp::complex<float>, devMem>& source,
-		      cusp::array1d<cusp::complex<float>, devMem>&
+		      cusp::array1d<cusp::complex<double>, devMem>& source,
+		      cusp::array1d<cusp::complex<double>, devMem>&
 		      tempSource);
 
     void bicgstab(const complexHybridHost& hostDirac,
 		  const complexHybridHost& hostSourceSmear,
 		  const complexHybridHost& hostSinkSmear,
 		  const int spatialIndex,
-		  cusp::array2d<cusp::complex<float>, hostMem>& propagator,
+		  cusp::array2d<cusp::complex<double>, hostMem>& propagator,
 		  const int verbosity);
 
     void cg(const complexHybridHost& hostDiracDiracAdjoint,
@@ -41,7 +41,7 @@ namespace pyQCD
 	    const complexHybridHost& hostSourceSmear,
 	    const complexHybridHost& hostSinkSmear,
 	    const int spatialIndex,
-	    cusp::array2d<cusp::complex<float>, hostMem>& propagator,
+	    cusp::array2d<cusp::complex<double>, hostMem>& propagator,
 	    const int verbosity);
   }
 }
