+++ pyQCD/core/src/cuda_utils.cu
@@ -6,13 +6,13 @@ namespace pyQCD
   {  
     void createSource(const int spatialIndex, const int spin, const int colour,
 		      const complexHybridDev& smearingOperator,
-		      cusp::array1d<cusp::complex<float>, devMem>& source,
-		      cusp::array1d<cusp::complex<float>, devMem>&
+		      cusp::array1d<cusp::complex<double>, devMem>& source,
+		      cusp::array1d<cusp::complex<double>, devMem>&
 		      tempSource)
     {
       int index = colour + 3 * (spin + spatialIndex);
-      cusp::blas::fill(tempSource, cusp::complex<float>(0.0, 0.0));
-      tempSource[index] = cusp::complex<float>(1.0, 0.0);
+      cusp::blas::fill(tempSource, cusp::complex<double>(0.0, 0.0));
+      tempSource[index] = cusp::complex<double>(1.0, 0.0);
 
       cusp::multiply(smearingOperator, tempSource, source);
     }
@@ -23,7 +23,7 @@ namespace pyQCD
 		  const complexHybridHost& hostSourceSmear,
 		  const complexHybridHost& hostSinkSmear,
 		  const int spatialIndex,
-		  cusp::array2d<cusp::complex<float>, hostMem>& propagator,
+		  cusp::array2d<cusp::complex<double>, hostMem>& propagator,
 		  const int verbosity)
     {
       // Get the size of the Dirac matrix
@@ -38,24 +38,24 @@ namespace pyQCD
 	std::cout << " Done!" << std::endl;
 
       // Create some device arrays to hold the source and solution
-      cusp::array1d<cusp::complex<float>, devMem>
-	source(nCols, cusp::complex<float>(0, 0));
-      cusp::array1d<cusp::complex<float>, devMem>
-	solution(nCols, cusp::complex<float>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	source(nCols, cusp::complex<double>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	solution(nCols, cusp::complex<double>(0, 0));
 
       // And another load as we'll need duplicates for the
       // multiplication routines
-      cusp::array1d<cusp::complex<float>, devMem>
-	tempSource(nCols, cusp::complex<float>(0, 0));
-      cusp::array1d<cusp::complex<float>, devMem>
-	tempSolution(nCols, cusp::complex<float>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	tempSource(nCols, cusp::complex<double>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	tempSolution(nCols, cusp::complex<double>(0, 0));
 
       // Create a temporary propagator variable for fast access
-      cusp::array2d<cusp::complex<float>, devMem>
-	tempPropagator(nCols, 12, cusp::complex<float>(0, 0));
+      cusp::array2d<cusp::complex<double>, devMem>
+	tempPropagator(nCols, 12, cusp::complex<double>(0, 0));
 
       // Create the preconditioner
-      cusp::precond::diagonal<cusp::complex<float>, devMem>
+      cusp::precond::diagonal<cusp::complex<double>, devMem>
 	preconditioner(devDirac);
       
       // Loop through all spins and colours and do the inversions
@@ -68,7 +68,7 @@ namespace pyQCD
 	  createSource(spatialIndex, i, j, devSourceSmear, source,
 		       tempSource);
 	  // Set up the monitor for use in the solver
-	  cusp::default_monitor<cusp::complex<float> >
+	  cusp::default_monitor<cusp::complex<double> >
 	    monitor(source, 1000, 0, 1e-8);
 	  
 	  // Do the inversion
@@ -77,7 +77,7 @@ namespace pyQCD
 	  // Smear at the sink
 	  cusp::multiply(devSinkSmear, solution, tempSolution);
 	  // Create a view to the relevant column of the propagator
-	  cusp::array2d<cusp::complex<float>, devMem>::column_view
+	  cusp::array2d<cusp::complex<double>, devMem>::column_view
 	    propagatorView = tempPropagator.column(j + 3 * i);
 	  // Copy the solution to the propagator output
 	  cusp::copy(tempSolution, propagatorView);
@@ -104,7 +104,7 @@ namespace pyQCD
 	    const complexHybridHost& hostSourceSmear,
 	    const complexHybridHost& hostSinkSmear,
 	    const int spatialIndex,
-	    cusp::array2d<cusp::complex<float>, hostMem>& propagator,
+	    cusp::array2d<cusp::complex<double>, hostMem>& propagator,
 	    const int verbosity)
     {
       // Get the size of the Dirac matrix
@@ -120,24 +120,24 @@ namespace pyQCD
 	std::cout << " Done!" << std::endl;
 
       // Create some device arrays to hold the source and solution
-      cusp::array1d<cusp::complex<float>, devMem>
-	source(nCols, cusp::complex<float>(0, 0));
-      cusp::array1d<cusp::complex<float>, devMem>
-	solution(nCols, cusp::complex<float>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	source(nCols, cusp::complex<double>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	solution(nCols, cusp::complex<double>(0, 0));
 
       // And another load as we'll need duplicates for the
       // multiplication routines
-      cusp::array1d<cusp::complex<float>, devMem>
-	tempSource(nCols, cusp::complex<float>(0, 0));
-      cusp::array1d<cusp::complex<float>, devMem>
-	tempSolution(nCols, cusp::complex<float>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	tempSource(nCols, cusp::complex<double>(0, 0));
+      cusp::array1d<cusp::complex<double>, devMem>
+	tempSolution(nCols, cusp::complex<double>(0, 0));
 
       // Create a temporary propagator variable for fast access
-      cusp::array2d<cusp::complex<float>, devMem>
-	tempPropagator(nCols, 12, cusp::complex<float>(0, 0));
+      cusp::array2d<cusp::complex<double>, devMem>
+	tempPropagator(nCols, 12, cusp::complex<double>(0, 0));
 
       // Create the preconditioner
-      cusp::precond::diagonal<cusp::complex<float>, devMem>
+      cusp::precond::diagonal<cusp::complex<double>, devMem>
 	preconditioner(devM);
 
       // Loop through all spins and colours and do the inversions
@@ -150,7 +150,7 @@ namespace pyQCD
 	  createSource(spatialIndex, i, j, devSourceSmear, source,
 		       tempSource);
 	  // Set up the monitor for use in the solver
-	  cusp::default_monitor<cusp::complex<float> >
+	  cusp::default_monitor<cusp::complex<double> >
 	    monitor(source, 1000, 0, 1e-8);
 	  
 	  // Do the inversion
@@ -161,7 +161,7 @@ namespace pyQCD
 	  
 	  cusp::multiply(devSinkSmear, tempSolution, solution);
 	  // Create a view to the relevant column of the propagator
-	  cusp::array2d<cusp::complex<float>, devMem>::column_view
+	  cusp::array2d<cusp::complex<double>, devMem>::column_view
 	    propagatorView = tempPropagator.column(j + 3 * i);
 	  // Copy the solution to the propagator output
 	  cusp::copy(solution, propagatorView);
